requirements.txt
numpy
pandas
matplotlib
plotly
tensorflow
yfinance
streamlit
telethon
python-dotenv
reportlab Pillow